TaskFlow/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
python-multipart
email-validator
python-dotenv
